Metrics Prom scrape annotation
The Service port is not used for scraping. The actual Endpoints resources
part of Service are at 9117 (targetPort/containerPort). The best practice
for scraping metrics is to do that against Pods and/or Endpoints, not Services
ports directly. It is known to cause issues if more than one endpoint is behind
the service, as metric queries will be load-balanced as well leading to
inaccurate histograms.